requirements.txt
requests
beautifulsoup4
supabase
python-dotenv
playwright
transformers
torch
tensorflow
transformers
torch
datasets
